fix: sort order with different values
